Fix comments from PR
