+++ helpdesk-bot/src/main/java/org/symphonyoss/symphony/bots/helpdesk/bot/HelpDeskServiceConfiguration.java
@@ -78,7 +78,7 @@ public class HelpDeskServiceConfiguration {
 
     AgentExternalCheck agentExternalCheck =
         new AgentExternalCheck(configuration.getHelpDeskBotUrl(),
-            configuration.getHelpDeskServiceUrl(), configuration.getGroupId(), ticketClient);
+            configuration.getHelpDeskServiceUrl(), configuration.getGroupId(), ticketClient, symphonyClient);
 
     agentMakerCheckerService.addCheck(agentExternalCheck);
