+++ helpdesk-maker-checker-service/src/main/java/org/symphonyoss/symphony/bots/helpdesk/makerchecker/model/check/AgentExternalCheck.java
@@ -4,8 +4,11 @@ import static org.symphonyoss.symphony.bots.helpdesk.service.ticket.client.Ticke
     .TicketStateType.UNRESOLVED;
 
 import org.apache.commons.codec.binary.Base64;
+import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.apache.commons.lang3.StringUtils;
+import org.symphonyoss.client.SymphonyClient;
+import org.symphonyoss.client.exceptions.AttachmentsException;
 import org.symphonyoss.symphony.bots.helpdesk.makerchecker.message.MakerCheckerMessageBuilder;
 import org.symphonyoss.symphony.bots.helpdesk.makerchecker.model.AttachmentMakerCheckerMessage;
 import org.symphonyoss.symphony.bots.helpdesk.makerchecker.model.MakerCheckerMessage;
@@ -13,12 +16,19 @@ import org.symphonyoss.symphony.bots.helpdesk.service.model.Ticket;
 import org.symphonyoss.symphony.bots.helpdesk.service.ticket.client.TicketClient;
 import org.symphonyoss.symphony.clients.model.SymAttachmentInfo;
 import org.symphonyoss.symphony.clients.model.SymMessage;
+import org.symphonyoss.symphony.clients.model.SymStream;
 
+import java.io.ByteArrayInputStream;
+import java.io.File;
+import java.io.IOException;
+import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
+
+import javax.ws.rs.BadRequestException;
 
 /**
  * Created by nick.tarsillo on 10/20/17.
@@ -26,6 +36,9 @@ import java.util.Set;
 public class AgentExternalCheck implements Checker {
 
   private static final int MAKERCHECKER_ID_LENGTH = 10;
+  private static final String MESSAGE_COULD_NOT_CREATE_TEMP_FILE = "Couldn't create a temp file.";
+  private static final String MESSAGE_ATTACHMENT_NOT_FOUND = "Attachment not found.";
+  private static final String MESSAGE_FAILED_TO_CREATE_FILE = "Failed to create File";
 
   private final String ATTACHMENT = "ATTACHMENT";
 
@@ -37,12 +50,15 @@ public class AgentExternalCheck implements Checker {
 
   private final TicketClient ticketClient;
 
+  private final SymphonyClient symphonyClient;
+
   public AgentExternalCheck(String botHost, String serviceHost, String groupId,
-      TicketClient ticketClient) {
+      TicketClient ticketClient, SymphonyClient symphonyClient) {
     this.botHost = botHost;
     this.serviceHost = serviceHost;
     this.groupId = groupId;
     this.ticketClient = ticketClient;
+    this.symphonyClient = symphonyClient;
   }
 
   @Override
@@ -129,5 +145,66 @@ public class AgentExternalCheck implements Checker {
     return StringUtils.isNotBlank(makerCheckerMessage.getType()) &&
         makerCheckerMessage.getType().equals(ATTACHMENT);
   }
+
+  @Override
+  public Set<SymMessage> makeApprovedMessages(MakerCheckerMessage makerCheckerMessage, SymMessage symMessage) {
+    Set<SymMessage> symApprovedMessages = new HashSet<>();
+
+    for(String streamId: makerCheckerMessage.getProxyToStreamIds()) {
+      SymMessage approvedMessage = new SymMessage();
+
+      SymStream stream = new SymStream();
+      stream.setStreamId(streamId);
+      approvedMessage.setStreamId(streamId);
+
+      approvedMessage.setStream(stream);
+      approvedMessage.setMessage(symMessage.getMessage());
+      approvedMessage.setEntityData(symMessage.getEntityData());
+      approvedMessage.setTimestamp(symMessage.getTimestamp());
+      approvedMessage.setFromUserId(symMessage.getFromUserId());
+
+      SymAttachmentInfo symAttachmentInfo = getApprovedAttachment(makerCheckerMessage, symMessage);
+      if (symAttachmentInfo != null && symAttachmentInfo.getId() != null) {
+        List<SymAttachmentInfo> attachmentInfoList = new ArrayList<>();
+        attachmentInfoList.add(symAttachmentInfo);
+        approvedMessage.setAttachments(attachmentInfoList);
+
+        File file = getFileAttachment(attachmentInfoList.get(0), symMessage);
+        approvedMessage.setAttachment(file);
+      }
+
+      symApprovedMessages.add(approvedMessage);
+    }
+
+    return symApprovedMessages;
+  }
+
+  private File getFileAttachment(SymAttachmentInfo symAttachmentInfo, SymMessage symMessage) {
+    File tempFile;
+    try {
+      String prefix = symAttachmentInfo.getName().split("\\.")[0];
+      String suffix = "." + symAttachmentInfo.getName().split("\\.")[1];
+      tempFile = File.createTempFile(prefix, suffix);
+    } catch (IOException e) {
+      throw new BadRequestException(MESSAGE_COULD_NOT_CREATE_TEMP_FILE);
+    }
+
+    byte[] aByte;
+    try {
+      aByte = symphonyClient.getAttachmentsClient().getAttachmentData(symAttachmentInfo, symMessage);
+    } catch (AttachmentsException e) {
+      throw new BadRequestException(MESSAGE_ATTACHMENT_NOT_FOUND);
+    }
+
+    InputStream inputStream = new ByteArrayInputStream(aByte);
+    try {
+      FileUtils.copyInputStreamToFile(inputStream, tempFile);
+    } catch (IOException e) {
+      throw new BadRequestException(MESSAGE_FAILED_TO_CREATE_FILE);
+    }
+
+    return tempFile;
+  }
+
 
 }
