+++ helpdesk-maker-checker-service/src/main/java/org/symphonyoss/symphony/bots/helpdesk/makerchecker/model/check/Checker.java
@@ -30,4 +30,6 @@ public interface Checker {
 
   boolean isCheckerType(MakerCheckerMessage makerCheckerMessage);
 
+  Set<SymMessage> makeApprovedMessages(MakerCheckerMessage makerCheckerMessage, SymMessage symMessage);
+
 }
